requirements.txt
faiss-cpu==1.9.0.post1
langchain_community==0.3.9
langchain_unstructured==0.1.6
langchain_huggingface==0.1.2
streamlit==1.48.1
google-api-python-client==2.161.0
unstructured==0.18.13
